+++ backend/common/src/stats.rs
@@ -1,4 +1,4 @@
 pub struct Stats {
-    pub event_count: i32,
-    pub person_count: i32,
+    pub event_count: i64,
+    pub person_count: i64,
 }
